ContactPage update
